# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "InstaClone API Documentation",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "Servidor de API Principal"
        }
    ],
    "paths": {
...
        "/feed": {
            "get": {
                "tags": [
                    "Feed"
                ],
                "summary": "Obter feed de postagens",
                "description": "Retorna os posts paginados dos usuários seguidos e do próprio usuário logado. Requer autenticação.",
                "operationId": "getFeed",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Número da página (padrão: 1)",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1,
                            "minimum": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Quantidade de posts por página (padrão: 15)",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 15,
                            "maximum": 100,
                            "minimum": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Feed retornado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "current_page": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "user_id": {
                                                        "type": "integer",
                                                        "example": 5
                                                    },
                                                    "caption": {
                                                        "type": "string",
                                                        "example": "Dia lindo no parque! 🌞"
                                                    },
                                                    "image_url": {
                                                        "type": "string",
                                                        "example": "/storage/posts/image_1.jpg"
                                                    },
                                                    "likes_count": {
                                                        "type": "integer",
                                                        "example": 42
                                                    },
                                                    "comments_count": {
                                                        "type": "integer",
                                                        "example": 8
                                                    },
                                                    "is_liked": {
                                                        "type": "boolean",
                                                        "example": false
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "example": "2026-05-22T10:30:00Z"
                                                    },
                                                    "user": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "integer",
                                                                "example": 5
                                                            },
                                                            "name": {
                                                                "type": "string",
                                                                "example": "João Silva"
                                                            },
                                                            "username": {
                                                                "type": "string",
                                                                "example": "joao_silva"
                                                            },
                                                            "avatar": {
                                                                "type": "string",
                                                                "example": "/storage/avatars/joao.jpg"
                                                            },
                                                            "is_following": {
                                                                "type": "boolean",
                                                                "example": true
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "first_page_url": {
                                            "type": "string",
                                            "example": "/api/feed?page=1"
                                        },
                                        "from": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "last_page": {
                                            "type": "integer",
                                            "example": 5
                                        },
                                        "last_page_url": {
                                            "type": "string",
                                            "example": "/api/feed?page=5"
                                        },
                                        "next_page_url": {
                                            "type": "string",
                                            "example": "/api/feed?page=2"
                                        },
                                        "path": {
                                            "type": "string",
                                            "example": "/api/feed"
                                        },
                                        "per_page": {
                                            "type": "integer",
                                            "example": 15
                                        },
                                        "prev_page_url": {
                                            "type": "string",
                                            "example": null
                                        },
                                        "to": {
                                            "type": "integer",
                                            "example": 15
                                        },
                                        "total": {
                                            "type": "integer",
                                            "example": 75
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autenticado"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}